chore(media): move utility method outside
