+++ src/utils/media.ts
@@ -1,4 +1,5 @@
-import { MediaLabels } from "types/media"
+import { MediaData } from "types/directory"
+import { MediaLabels, SelectedMediaDto } from "types/media"
 
 // Utility method to help ease over the various labels associated
 // with the media type so that we can avoid repeated conditionals
@@ -20,4 +21,15 @@ export const getMediaLabels = (mediaType: "files" | "images"): MediaLabels => {
     singularDirectoryLabel: "album",
     pluralDirectoryLabel: "albums",
   }
+}
+
+// Utility method to construct a SelectedMediaDto from MediaData
+export const getSelectedMediaDto = (fileData: MediaData) => {
+  const selectedData: SelectedMediaDto = {
+    filePath: fileData.mediaPath,
+    sha: fileData.sha,
+    size: fileData.size || 0,
+  }
+
+  return selectedData
 }
